requirements.txt
# ============================
# Core Numerical Stack
# ============================
numpy==1.26.4
pandas==2.2.1
scipy==1.12.0

# ============================
# Visualization
# ============================
matplotlib==3.8.3

# ============================
# Machine Learning
# ============================
scikit-learn==1.4.2
xgboost==2.0.3
lightgbm==4.3.0
shap==0.44.1

# ============================
# Time Series
# ============================
prophet==1.1.5

# ============================
# Utilities
# ============================
joblib==1.3.2
tqdm==4.66.2
python-dotenv==1.0.1
pyyaml==6.0.1

# ============================
# LLM / GenAI
# ============================
openai==1.12.0
langchain==0.1.13
tiktoken==0.6.0
